requirements.txt
# PyTorch extra index for CUDA wheels
--extra-index-url https://download.pytorch.org/whl/cu121

numpy<2
python-dotenv

torch
torch
torchvision
torchaudio

# IMPORTANT: torchtext==0.15.2 has a HARD dependency: torch==2.0.1 This is REQUIRED for the custom from-scratch Transformer codebase. Upgrading torch beyond 2.0.1 WILL break torchtext 0.15.2.
torchtext==0.15.2


# Required for using HF pretrained models
# python -m pip install --upgrade --force-reinstall torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
transformers==4.33.3

sentencepiece # Requiredfor Marian / OPUS-MT tokenizers
sacremoses # Removes MarianTokenizer warning

datasets==2.15.0
tokenizers==0.13.3

torchmetrics==1.0.3
tensorboard==2.13.0
altair==5.1.1
wandb==0.15.9
huggingface_hub>=0.18.0

# OpenAI SDK (for LLM-based translation)
openai>=1.0.0

# Cuda sanity check
# python -c "import torch; print('torch:', torch.__version__); print('cuda:', torch.cuda.is_available())"
